decompress: move from arm9 to common
